fixed the individual game update
the problem was that the individualmlbgame function was being repeatedly called but it was empty so it wasnt returning new data. with the line added to the very top of the function it defaults the parameter to the selectedGameId which is passed into the function. this allows the function to call the api and get the proper game data every time.