updated individual view
updated the individual mlb game view. many formatting updates like not showing balls strikes and outs when its at the end or middle of an inning. added which inning it is. added the batter and pitcher matchup. changed the current matchup to not show before a pitch is thrown to not have redundant info showing.