updated the 3 files above
updated script.js to remove the nfl variables and put them in the new nflVariables file. also added scheduled game html code to a function in the nfl variables file. now when the script file wants to show a scheduled nfl game it just calls the scheduledGame function. I did this because it was being written out in both that and show all. I will update all of the rest of code base to reflect that change.