updated the 3 above files
mlbVariables: moved the game status html to this file from the main one so that the logic can be carried over for the short list and all games.
nflVariables: did the same for this file and made sure to update the links in these files from mlb. and nfl. to this.
script: updated points where it checks the game status and calls appropriate function from its respective variables file.