updated mlb and nflVariables.js and script.js
script.js - added nfl individual game function. working properly right now but will need to test further when there is an active game.
nflVariables - added gameStatus so that the new renderIndividualView function can take that string and send the html to the script file and display information for that game
mlbVariables - also added the same renderIndividualView to display its own relevant info. this cleared all of the logic out of the mlbindividual function so that its cleaner