updated mlbVariables file
updated the html that is being brought in. now on the "mini" views inprogress scheduled and final it is structured and styled with the other update for it to look good. next I will be updating the individual views to keep that nice view over the whole site. from there the rest of the leagues will be updated as well.