updated styles
updated the game row hover to a hoverer class. this class now surrounds every brought in function to show game details. this fix was made because on hover of the base game row it was transforming every second when the api call is made. so now the hover is on the container around each game and works as it should.