updated mlb variables
updated mlbvariables file with logic for weather not being available, future weather is a number, and the display properties along with that. also for mobile view added a small logo next to the probable pitcher so that people know who plays for who when they arent in the same column as the teams logo.